Moved demo to it's own package to test the customizable resources.
We're now leaving the userStyleSheet un-set so that the user can set
their own style without breaking things. This means that the stylesheets
have to be referenced in the HTML; this means that (for now) we can only
use the default HTML—no HTML files loaded from outside. The reason for
this is using a HTML file from outside will break the references to
the CSS files. A work-around to this is to extract the contents of the
CSS files into style tags.